requirements.txt
absl-py==1.4.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
Babel==2.10.3
backcall==0.2.0
bleach==5.0.1
branca==0.5.0
cachetools==5.3.0
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cmake==3.25.0
cycler==0.11.0
debugpy==1.6.3
deepface==0.0.78
defusedxml==0.7.1
dlib==19.24.0
executing==1.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastjsonschema==2.16.1
filelock==3.9.0
fire==0.5.0
flatbuffers==23.1.21
folium==0.12.1.post1
fonttools==4.37.1
gast==0.4.0
gdown==4.6.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.51.1
h5py==3.8.0
idna==3.4
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.10
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.2
mistune==2.0.4
mtcnn==0.1.1
nbclassic==0.4.3
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pickleshare==0.7.5
Pillow==9.4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
PythonTurtle==0.3.2
pytz==2022.7.1
pyzmq==23.2.1
requests==2.28.2
requests-oauthlib==1.3.1
retina-face==0.0.13
rsa==4.9
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.5.0
statsmodels==0.13.2
tenacity==8.0.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tqdm==4.64.1
traitlets==5.3.0
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
wrapt==1.14.1
